refactor(editor): Move serializeNode into nodeTransforms as standalone function
