+++ packages/frontend/editor-ui/src/app/stores/workflowDocument.store.ts
@@ -32,7 +32,7 @@ import { useUIStore } from '@/app/stores/ui.store';
 import { useNodeTypesStore } from '@/app/stores/nodeTypes.store';
 import { useWorkflowsStore } from '@/app/stores/workflows.store';
 import { useNodeHelpers } from '@/app/composables/useNodeHelpers';
-import { useWorkflowDocumentNodeSerialization } from './workflowDocument/useWorkflowDocumentNodeSerialization';
+import { serializeNode } from '@/app/utils/nodes/nodeTransforms';
 import type { WorkflowObjectAccessors } from '../types';
 import type { IWorkflowDb } from '@/Interface';
 import type { INode, IPinData } from 'n8n-workflow';
@@ -79,7 +79,6 @@ type MetaReturn = ReturnType<typeof useWorkflowDocumentMeta>;
 type PinDataReturn = ReturnType<typeof useWorkflowDocumentPinData>;
 type SettingsReturn = ReturnType<typeof useWorkflowDocumentSettings>;
 type NodeMetadataReturn = ReturnType<typeof useWorkflowDocumentNodeMetadata>;
-type NodeSerializationReturn = ReturnType<typeof useWorkflowDocumentNodeSerialization>;
 
 // Pairwise collision checks — add new composables here when they are created.
 // If any pair shares a key, the corresponding tuple slot becomes an error type
@@ -102,11 +101,6 @@ void (0 as unknown as [
 	AssertNoOverlap<NodeMetadataReturn, PinDataReturn>,
 	AssertNoOverlap<NodeMetadataReturn, MetaReturn>,
 	AssertNoOverlap<NodeMetadataReturn, SettingsReturn>,
-	AssertNoOverlap<NodeSerializationReturn, NodesReturn>,
-	AssertNoOverlap<NodeSerializationReturn, ConnectionsReturn>,
-	AssertNoOverlap<NodeSerializationReturn, GraphReturn>,
-	AssertNoOverlap<NodeSerializationReturn, ExpressionReturn>,
-	AssertNoOverlap<NodeSerializationReturn, NodeMetadataReturn>,
 ]);
 
 export type WorkflowDocumentId = `${string}@${string}`;
@@ -188,9 +182,6 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 			});
 		const workflowDocumentGraph = useWorkflowDocumentGraph(workflowObject);
 		const workflowDocumentExpression = useWorkflowDocumentExpression(workflowObject);
-		const workflowDocumentNodeSerialization = useWorkflowDocumentNodeSerialization({
-			getNodeType: (typeName, version) => nodeTypesStore.getNodeType(typeName, version),
-		});
 
 		// --- Cross-cut orchestration ---
 		// Each composable is self-contained and unaware of its siblings. This
@@ -208,8 +199,8 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 		}
 
 		function serialize(): WorkflowData {
-			const nodes: INode[] = workflowDocumentNodes.allNodes.value.map(
-				workflowDocumentNodeSerialization.serializeNode,
+			const nodes: INode[] = workflowDocumentNodes.allNodes.value.map((node) =>
+				serializeNode(nodeTypesStore, node),
 			);
 
 			// Deep-copy connections to create an in-time snapshot consistent with nodes.
@@ -364,7 +355,6 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 			...workflowDocumentGraph,
 			...workflowDocumentExpression,
 			...workflowDocumentNodeMetadata,
-			...workflowDocumentNodeSerialization,
 			removeAllNodes,
 			hydrate,
 			reset,
